refactor: folder organization
